Проводит декомпозицию презентера
