+++ src/framework/observable.js
@@ -26,8 +26,8 @@ export default class Observable {
    * @param {*} event Тип события
    * @param {*} payload Дополнительная информация
    */
-  _notify(event, payload) {
-    this.#observers.forEach((observer) => observer(event, payload));
+  _notify(updateType, updatedFilm) {
+    this.#observers.forEach((observer) => observer(updateType, updatedFilm));
   }
 }
