+++ src/model/films-model.js
@@ -11,8 +11,8 @@ export default class FilmsModel extends Observable {
     return this.#films;
   }
 
-  updateFilm = (updateType, update) => {
-    const index = this.#films.findIndex((film) => film.id === update.id);
+  updateFilm = (updateType, updatedFilm) => {
+    const index = this.#films.findIndex((film) => film.id === updatedFilm.id);
 
     if (index === -1) {
       throw new Error('Can\'t update unexisting film');
@@ -20,10 +20,10 @@ export default class FilmsModel extends Observable {
 
     this.#films = [
       ...this.#films.slice(0, index),
-      update,
+      updatedFilm,
       ...this.#films.slice(index + 1),
     ];
 
-    this._notify(updateType, update);
+    this._notify(updateType, updatedFilm);
   };
 }
