+++ src/presenter/film-presenter.js
@@ -1,22 +1,22 @@
 import FilmCardView from '../view/film-card-view.js';
-import FilmPopupView from '../view/film-popup-view.js';
-import { UserAction, UpdateType } from '../const.js';
-import { render, remove, replace, RenderPosition } from '../framework/render.js';
+import FilmPopupPresenter from './film-popup-presenter.js';
+import { UpdateType } from '../const.js';
+import { render, remove, replace } from '../framework/render.js';
 
 export default class FilmPresenter {
   #film = null;
   #comments = null;
+  #filmsModel = null;
   #changeData = null;
-  #closeAnyOpenPopup = null;
   #filmCardComponent = null;
-  #filmPopupComponent = null;
   #currentFilmsContainer = null;
+  #filmPopupPresenter = new FilmPopupPresenter();
 
-  constructor(comments, changeData, closeAnyOpenPopup, container) {
+  constructor(comments, changeData, container, filmsModel) {
     this.#comments = comments;
+    this.#filmsModel = filmsModel;
     this.#changeData = changeData;
     this.#currentFilmsContainer = container;
-    this.#closeAnyOpenPopup = closeAnyOpenPopup;
   }
 
   init = (film) => {
@@ -25,7 +25,7 @@ export default class FilmPresenter {
     const prevFilmComponent = this.#filmCardComponent;
 
     this.#filmCardComponent = new FilmCardView(this.#film);
-    this.#filmCardComponent.setClickHandler(() => this.#addFilmPopup(this.#film, this.#comments));
+    this.#filmCardComponent.setClickHandler(() => this.#filmPopupPresenter.init(this.#film, this.#comments, this.#changeData, this.#filmsModel));
     this.#filmCardComponent.setFavoriteClickHandler(this.#handleFavoriteClick);
     this.#filmCardComponent.setWatchedClickHandler(this.#handleWatchedClick);
     this.#filmCardComponent.setWatchlistClickHandler(this.#handleWatchlistClick);
@@ -37,10 +37,6 @@ export default class FilmPresenter {
 
     if (this.#currentFilmsContainer.contains(prevFilmComponent.element)) {
       replace(this.#filmCardComponent, prevFilmComponent);
-
-      if (document.querySelector('.film-details')) {
-        this.#addFilmPopup(this.#film, this.#comments);
-      }
     }
 
     remove(prevFilmComponent);
@@ -50,44 +46,9 @@ export default class FilmPresenter {
     remove(this.#filmCardComponent);
   };
 
-  #addFilmPopup = (film, commentsList) => {
-    this.#closeAnyOpenPopup();
-
-    const siteFooterElement = document.querySelector('.footer');
-    const selectedComments = commentsList.filter(({ id }) => film.comments.some((commentId) => commentId === Number(id)));
-    this.#filmPopupComponent = new FilmPopupView(film, selectedComments);
-
-    render(this.#filmPopupComponent, siteFooterElement, RenderPosition.AFTEREND);
-    document.body.classList.add('hide-overflow');
-
-    document.addEventListener('keydown', this.#onDocumentKeyDown);
-    this.#filmPopupComponent.setClickHandler(this.#onCloseBtnClick);
-    this.#filmPopupComponent.setFavoriteClickHandler(this.#handleFavoriteClick);
-    this.#filmPopupComponent.setWatchedClickHandler(this.#handleWatchedClick);
-    this.#filmPopupComponent.setWatchlistClickHandler(this.#handleWatchlistClick);
-  };
-
-  removeFilmPopup = () => {
-    document.body.classList.remove('hide-overflow');
-    remove(this.#filmPopupComponent);
-    document.removeEventListener('keydown', this.#onDocumentKeyDown);
-  };
-
-  #onCloseBtnClick = () => {
-    this.removeFilmPopup();
-  };
-
-  #onDocumentKeyDown = (evt) => {
-    if (evt.code === 'Escape') {
-      evt.preventDefault();
-      this.removeFilmPopup();
-    }
-  };
-
   #handleFavoriteClick = () => {
     this.#changeData(
-      UserAction.UPDATE_FILM,
-      UpdateType.MINOR,
+      UpdateType.PATCH,
       {...this.#film,
         userDetails: {
           favorite: !this.#film.userDetails.favorite,
@@ -100,8 +61,7 @@ export default class FilmPresenter {
 
   #handleWatchedClick = () => {
     this.#changeData(
-      UserAction.UPDATE_FILM,
-      UpdateType.MINOR,
+      UpdateType.PATCH,
       {...this.#film,
         userDetails: {
           favorite: this.#film.userDetails.favorite,
@@ -114,8 +74,7 @@ export default class FilmPresenter {
 
   #handleWatchlistClick = () => {
     this.#changeData(
-      UserAction.UPDATE_FILM,
-      UpdateType.MINOR,
+      UpdateType.PATCH,
       {...this.#film,
         userDetails: {
           favorite: this.#film.userDetails.favorite,
